Introduce a new module `io.github.linghengqian:hive-server2-jdbc-driver-reachability-metadata` to store GraalVM Reachability Metadata for third-party dependencies
