--- hive-server2-jdbc-driver-uber/src/main/resources/META-INF/native-image/com.github.docker-java/docker-java-api/3.4.2/proxy-config.json
@@ -1,6 +1,0 @@
-[
-  {
-    "condition":{"typeReachable":"com.github.dockerjava.transport.NamedPipeSocket"},
-    "interfaces":["com.github.dockerjava.transport.NamedPipeSocket$Kernel32"]
-  }
-]
