--- hive-server2-jdbc-driver-uber/src/main/resources/META-INF/native-image/com.github.docker-java/docker-java-api/3.4.2/resource-config.json
@@ -1,8 +1,0 @@
-{
-  "resources":{
-  "includes":[{
-    "condition":{"typeReachable":"com.github.dockerjava.zerodep.shaded.org.apache.hc.core5.util.VersionInfo"},
-    "pattern":"\\Qcom/github/dockerjava/zerodep/shaded/org/apache/hc/client5/version.properties\\E"
-  }]},
-  "bundles":[]
-}
