--- hive-server2-jdbc-driver-uber/src/main/resources/META-INF/native-image/org.testcontainers/junit-jupiter/1.21.3/reflect-config.json
@@ -1,151 +1,0 @@
-[
-{
-  "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-  "name":"org.testcontainers.containers.ClickHouseProvider"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-  "name":"org.testcontainers.containers.MSSQLServerContainerProvider"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-  "name":"org.testcontainers.containers.PgVectorContainerProvider"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-  "name":"org.testcontainers.containers.PostgisContainerProvider"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-  "name":"org.testcontainers.containers.PostgreSQLContainerProvider"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-  "name":"org.testcontainers.containers.TimescaleDBContainerProvider"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.DockerClientFactory"},
-  "name":"org.testcontainers.dockerclient.DockerDesktopClientProviderStrategy"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.DockerClientFactory"},
-  "name":"org.testcontainers.dockerclient.DockerMachineClientProviderStrategy"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.DockerClientFactory"},
-  "name":"org.testcontainers.dockerclient.EnvironmentAndSystemPropertyClientProviderStrategy"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.DockerClientFactory"},
-  "name":"org.testcontainers.dockerclient.NpipeSocketClientProviderStrategy"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.DockerClientFactory"},
-  "name":"org.testcontainers.dockerclient.RootlessDockerClientProviderStrategy"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.DockerClientFactory"},
-  "name":"org.testcontainers.dockerclient.TestcontainersHostPropertyClientProviderStrategy"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.fasterxml.jackson.databind.ext.Java7Support"},
-  "name":"org.testcontainers.shaded.com.fasterxml.jackson.databind.ext.Java7SupportImpl",
-  "methods":[{"name":"<init>","parameterTypes":[] }]
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.dockerclient.DockerClientProviderStrategy"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.DockerConfigFile",
-  "queryAllDeclaredMethods":true,
-  "queryAllDeclaredConstructors":true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.dockerclient.EnvironmentAndSystemPropertyClientProviderStrategy"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.DockerConfigFile"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.fasterxml.jackson.databind.ObjectMapper"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.DockerConfigFile",
-  "allDeclaredFields":true,
-  "methods":[{"name":"<init>","parameterTypes":[] }, {"name":"setAuths","parameterTypes":["java.util.Map"] }]
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.DefaultDockerClientConfig$Builder"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.DockerConfigFile"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.DockerConfigFile"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.DockerConfigFile"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.DefaultInvocationBuilder"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.command.AbstrAsyncDockerCmd",
-  "queryAllDeclaredMethods":true,
-  "allDeclaredFields": true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.exec.CreateNetworkCmdExec"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.command.AbstrDockerCmd",
-  "queryAllDeclaredMethods":true,
-  "allDeclaredFields": true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.exec.CreateNetworkCmdExec"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.command.CreateNetworkCmdImpl",
-  "allDeclaredFields":true,
-  "queryAllDeclaredMethods":true,
-  "queryAllDeclaredConstructors":true,
-  "methods":[{"name":"getAttachable","parameterTypes":[] }, {"name":"getCheckDuplicate","parameterTypes":[] }, {"name":"getDriver","parameterTypes":[] }, {"name":"getEnableIPv6","parameterTypes":[] }, {"name":"getInternal","parameterTypes":[] }, {"name":"getIpam","parameterTypes":[] }, {"name":"getLabels","parameterTypes":[] }, {"name":"getName","parameterTypes":[] }, {"name":"getOptions","parameterTypes":[] }]
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.exec.ExecCreateCmdExec"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.command.ExecCreateCmdImpl",
-  "allDeclaredFields":true,
-  "queryAllDeclaredMethods":true,
-  "queryAllDeclaredConstructors":true,
-  "methods":[{"name":"getContainerId","parameterTypes":[] }, {"name":"getEnv","parameterTypes":[] }, {"name":"getPrivileged","parameterTypes":[] }, {"name":"getUser","parameterTypes":[] }, {"name":"getWorkingDir","parameterTypes":[] }]
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.fasterxml.jackson.databind.util.ClassUtil"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.command.CreateContainerCmdImpl",
-  "allDeclaredMethods": true,
-  "allDeclaredConstructors": true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.command.CreateContainerCmdImpl"},
-  "name":"org.testcontainers.shaded.com.github.dockerjava.core.command.CreateContainerCmdImpl$NetworkingConfig",
-  "allDeclaredFields": true,
-  "allDeclaredMethods": true,
-  "allDeclaredConstructors": true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.containers.GenericContainer"},
-  "name":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory$1",
-  "allDeclaredMethods": true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory"},
-  "name":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory$1"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory$1"},
-  "name":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory$1"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.org.hamcrest.TypeSafeMatcher"},
-  "name":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory$1"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.shaded.org.hamcrest.internal.ReflectiveTypeFinder"},
-  "name":"org.testcontainers.shaded.org.awaitility.core.ConditionFactory$1"
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.junit.jupiter.Testcontainers"},
-  "name":"org.testcontainers.junit.jupiter.Testcontainers",
-  "allPublicMethods": true
-},
-{
-  "condition":{"typeReachable":"org.testcontainers.junit.jupiter.Container"},
-  "name":"org.testcontainers.junit.jupiter.Container",
-  "allPublicMethods": true
-}
-]
