--- hive-server2-jdbc-driver-uber/src/main/resources/META-INF/native-image/org.testcontainers/junit-jupiter/1.21.3/resource-config.json
@@ -1,29 +1,0 @@
-{
-  "resources":{
-  "includes":[{
-    "condition":{"typeReachable":"org.testcontainers.jdbc.ContainerDatabaseDriver"},
-    "pattern":"\\QMETA-INF/services/org.testcontainers.containers.JdbcDatabaseContainerProvider\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.containers.GenericContainer"},
-    "pattern":"\\QMETA-INF/services/org.testcontainers.core.CreateContainerCmdModifier\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.utility.ImageNameSubstitutor"},
-    "pattern":"\\QMETA-INF/services/org.testcontainers.utility.ImageNameSubstitutor\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.shaded.com.google.common.io.Resources"},
-    "pattern":"\\Qcontainer-license-acceptance.txt\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.containers.wait.strategy.HostPortWaitStrategy"},
-    "pattern":"\\Qlogback-test.xml\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.utility.ClasspathScanner"},
-    "pattern":"\\Qtestcontainers.properties\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.shaded.com.github.dockerjava.core.DefaultDockerClientConfig"},
-    "pattern":"\\Qdocker-java.properties\\E"
-  }, {
-    "condition":{"typeReachable":"org.testcontainers.dockerclient.DockerClientProviderStrategy"},
-    "pattern":"\\QMETA-INF/services/java.net.spi.InetAddressResolverProvider\\E"
-  }]},
-  "bundles":[]
-}
