+++ reachability-metadata/src/main/resources/META-INF/native-image/ch.qos.logback/logback-classic/1.2.13/reflect-config.json
@@ -1,0 +1,53 @@
+[
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.joran.action.NestedComplexPropertyIA"},
+  "name":"ch.qos.logback.classic.encoder.PatternLayoutEncoder",
+  "methods":[{"name":"<init>","parameterTypes":[] }],
+  "queryAllPublicMethods":true
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.pattern.parser.Compiler"},
+  "name":"ch.qos.logback.classic.pattern.LevelConverter",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.pattern.parser.Compiler"},
+  "name":"ch.qos.logback.classic.pattern.LineSeparatorConverter",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.pattern.parser.Compiler"},
+  "name":"ch.qos.logback.classic.pattern.LoggerConverter",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.pattern.parser.Compiler"},
+  "name":"ch.qos.logback.classic.pattern.MessageConverter",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.pattern.parser.Compiler"},
+  "name":"ch.qos.logback.classic.pattern.ThreadConverter",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.joran.action.NestedComplexPropertyIA"},
+  "name":"ch.qos.logback.core.encoder.Encoder",
+  "methods":[{"name":"valueOf","parameterTypes":["java.lang.String"] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.joran.util.PropertySetter"},
+  "name":"ch.qos.logback.core.encoder.LayoutWrappingEncoder",
+  "methods":[{"name":"setParent","parameterTypes":["ch.qos.logback.core.spi.ContextAware"] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.joran.util.StringToObjectConverter"},
+  "name":"ch.qos.logback.core.spi.ContextAware",
+  "methods":[{"name":"valueOf","parameterTypes":["java.lang.String"] }]
+},
+{
+  "condition":{"typeReachable":"ch.qos.logback.core.joran.action.StatusListenerAction"},
+  "name":"ch.qos.logback.core.status.NopStatusListener",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+}
+]
