+++ reachability-metadata/src/main/resources/META-INF/native-image/com.github.docker-java/docker-java-api/3.4.2/proxy-config.json
