+++ reachability-metadata/src/main/resources/META-INF/native-image/io.github.linghengqian/hive-server2-jdbc-driver-reachability-metadata/native-image.properties
