+++ reachability-metadata/src/main/resources/META-INF/native-image/org.apache.groovy/groovy/4.0.22/resource-config.json
@@ -1,0 +1,11 @@
+{
+  "resources":{
+  "includes":[{
+    "condition":{"typeReachable":"org.codehaus.groovy.reflection.GeneratedMetaMethod$DgmMethodRecord"},
+    "pattern":"\\QMETA-INF/dgminfo\\E"
+  }, {
+    "condition":{"typeReachable":"org.codehaus.groovy.runtime.m12n.ExtensionModuleScanner"},
+    "pattern":"\\QMETA-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule\\E"
+  }]},
+  "bundles":[]
+}
