+++ reachability-metadata/src/main/resources/META-INF/native-image/org.junit.jupiter/junit-jupiter/5.13.4/reflect-config.json
@@ -1,0 +1,52 @@
+[
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.extension.RegisterExtension",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.condition.EnabledInNativeImage",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.condition.EnabledIfSystemProperty",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.extension.ExtendWith",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.Test",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.platform.commons.annotation.Testable",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.ReflectionUtils"},
+  "name":"org.junit.internal.AssumptionViolatedException",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.AutoClose",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.condition.DisabledOnOs",
+  "allPublicMethods": true
+},
+{
+  "condition":{"typeReachable":"org.junit.platform.commons.util.AnnotationUtils"},
+  "name":"org.junit.jupiter.api.Disabled",
+  "allPublicMethods": true
+}
+]
