+++ reachability-metadata/src/main/resources/META-INF/native-image/org.junit.jupiter/junit-jupiter/5.13.4/resource-config.json
@@ -1,0 +1,24 @@
+{
+  "resources":{
+  "includes":[{
+    "condition":{"typeReachable":"org.junit.platform.launcher.core.LauncherFactory"},
+    "pattern":"\\QMETA-INF/services/org.junit.platform.launcher.LauncherDiscoveryListener\\E"
+  }, {
+    "condition":{"typeReachable":"org.junit.platform.launcher.core.LauncherFactory"},
+    "pattern":"\\QMETA-INF/services/org.junit.platform.launcher.LauncherSessionListener\\E"
+  }, {
+    "condition":{"typeReachable":"org.junit.platform.launcher.core.LauncherFactory"},
+    "pattern":"\\QMETA-INF/services/org.junit.platform.launcher.PostDiscoveryFilter\\E"
+  }, {
+    "condition":{"typeReachable":"org.junit.platform.commons.util.ClasspathScannerLoader"},
+    "pattern":"\\QMETA-INF/services/org.junit.platform.commons.support.scanning.ClasspathScanner\\E"
+  }, {
+    "condition":{"typeReachable":"org.junit.platform.launcher.core.LauncherConfigurationParameters"},
+    "pattern":"\\Qjunit-platform.properties\\E"
+  }]},
+  "bundles":[
+    {
+      "name":"sun.util.logging.resources.logging"
+    }
+  ]
+}
