+++ reachability-metadata/src/main/resources/META-INF/native-image/org.testcontainers/junit-jupiter/1.21.3/resource-config.json
