+++ thin/src/test/java/io/github/linghengqian/hive/server2/jdbc/driver/thin/iceberg/AvroTest.java
