+++ thin/src/test/resources/META-INF/native-image/test-metadata/reflect-config.json
@@ -1,0 +1,74 @@
+[
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.iceberg.AvroTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.iceberg.AvroTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.iceberg.OrcTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.iceberg.OrcTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.iceberg.ParquetTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.iceberg.ParquetTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.AcidTableTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.AcidTableTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.ClassExistTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.ClassExistTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.InformationSchemaTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.InformationSchemaTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.StandaloneMetastoreTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.StandaloneMetastoreTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.SimpleTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.SimpleTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+},
+{
+  "condition":{"typeReachable":"io.github.linghengqian.hive.server2.jdbc.driver.thin.ZookeeperServiceDiscoveryTest"},
+  "name":"io.github.linghengqian.hive.server2.jdbc.driver.thin.ZookeeperServiceDiscoveryTest",
+  "allDeclaredConstructors": true,
+  "allDeclaredMethods": true,
+  "allPublicMethods": true,
+  "allDeclaredFields": true
+}
+]
