+++ thin/src/test/resources/information-schema/Dockerfile
