+++ uber/src/test/java/io/github/linghengqian/hive/server2/jdbc/driver/uber/ClassExistTest.java
